utils/requirements.txt
absl-py==1.2.0
snowflake-snowpark-python==0.11.0
databricks-sql-connector==2.1.0
google-cloud-bigquery==3.5.0
redshift_connector==2.0.910
pandas
